move bump utils to global module
