+++ llama-dev/llama_dev/pkg/bump.py
@@ -1,54 +1,13 @@
-import re
-from enum import Enum
-from pathlib import Path
-
 import click
-from packaging.version import Version
 
-from llama_dev.utils import find_all_packages, is_llama_index_package, load_pyproject
-
-
-class BumpType(str, Enum):
-    MAJOR = "major"
-    MINOR = "minor"
-    PATCH = "patch"
-
-
-def bump_version(current_version: str, bump_type: BumpType) -> str:
-    """Bump a version string according to semver rules."""
-    v = Version(current_version)
-
-    # Parse the version components
-    release = v.release
-    major = release[0] if len(release) > 0 else 0
-    minor = release[1] if len(release) > 1 else 0
-    micro = release[2] if len(release) > 2 else 0
-
-    version_str = ""
-    if bump_type == BumpType.MAJOR:
-        version_str = f"{major + 1}.0.0"
-    elif bump_type == BumpType.MINOR:
-        version_str = f"{major}.{minor + 1}.0"
-    elif bump_type == BumpType.PATCH:
-        version_str = f"{major}.{minor}.{micro + 1}"
-
-    return version_str
-
-
-def update_pyproject_version(package_path: Path, new_version: str) -> None:
-    """Update the version in a pyproject.toml file."""
-    pyproject_path = package_path / "pyproject.toml"
-
-    # Read the file content
-    with open(pyproject_path, "r") as f:
-        content = f.read()
-
-    pattern = r'(\[project\][^\[]*?version\s*=\s*["\'])([^"\']+)(["\'])'
-    new_content = re.sub(pattern, rf"\g<1>{new_version}\g<3>", content, flags=re.DOTALL)
-
-    # Write the updated content back
-    with open(pyproject_path, "w") as f:
-        f.write(new_content)
+from llama_dev.utils import (
+    BumpType,
+    bump_version,
+    find_all_packages,
+    is_llama_index_package,
+    load_pyproject,
+    update_pyproject_version,
+)
 
 
 @click.command(short_help="Bump package version")
